requirements.txt
appdirs==1.4.4
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==7.1.2
cryptography==39.0.0
Django==3.2.3
django-cors-headers==3.13.0
django-debug-toolbar==3.2.1
django-ninja==0.13.0
djangorestframework==3.12.4
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
jwt==1.3.1
mypy-extensions==0.4.3
mysqlclient==2.1.1
numpy==1.24.1
outcome==1.2.0
pandas==1.5.3
pathspec==0.8.1
pycparser==2.21
pydantic==1.8.2
PyMySQL==1.0.2
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
python-utils==3.4.5
pytz==2022.7.1
regex==2021.4.4
requests==2.28.2
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
toml==0.10.2
trio==0.22.0
trio-websocket==0.9.2
typing-extensions==3.10.0.0
urllib3==1.26.14
utils==1.0.1
wsproto==1.2.0
